pyproject.toml
[project]
name = "banking-app"
version = "0.1.0"
description = ""
authors = [
    {name = "Ullas Zingade",email = "ullas09zingade@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask (>=3.1.1,<4.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "mysql-connector-python (>=9.3.0,<10.0.0)",
    "pydantic (>=2.11.5,<3.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "flask-restful (>=0.3.10,<0.4.0)",
    "pymysql (>=1.1.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "banking_app", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
